irefactor: move gitnessclient module into gitprovider
